A moment's hesitation before... changing our approach on constants

+++ Push.Shopify/Api/Order/TaxLine.cs
@@ -3,7 +3,7 @@ namespace Push.Shopify.Api.Order
     public class TaxLine
     {
         public string price { get; set; }
-        public double rate { get; set; }
+        public decimal rate { get; set; }
         public string title { get; set; }
     }
 }
